% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_labels.R
\name{colnames_labels}
\alias{colnames_labels}
\title{Show labels of colnames}
\usage{
colnames_labels(df, order = FALSE)
}
\arguments{
\item{df}{one dataframe}

\item{order}{logical, whether order by colname}
}
\value{
a dataframe contains colname and label
}
\description{
Show labels of colnames
}
