% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_items}
\alias{config_items}
\title{Config items of 'NHANES' database}
\usage{
config_items(items)
}
\arguments{
\item{items}{items of 'NHANES' database. If items is missing, Demographics,
Dietary, Examination, Laboratory and Questionnaire will be used.}
}
\value{
add "items.nhanes" file to "nhanesR" directory
}
\description{
Config items of 'NHANES' database
}
\examples{
\donttest{
config_items(c("Demographics", "Dietary", "Examination",  "Laboratory", "Questionnaire"))
}
}
