% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_path}
\alias{config_path}
\title{Config path of 'NHANES' database}
\usage{
config_path(path)
}
\arguments{
\item{path}{path of 'NHANES' database}
}
\value{
change the internal path
}
\description{
Config path of 'NHANES' database
}
\examples{
\donttest{
config_path('f/nhanes')
}
}
