% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_temp}
\alias{config_temp}
\title{temporary config directory}
\usage{
config_temp()
}
\value{
temporary config directory
}
\description{
temporary config directory
}
\examples{
\donttest{
config_temp()
}
}
