% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_years}
\alias{config_years}
\title{Config years of 'NHANES' database}
\usage{
config_years(years)
}
\arguments{
\item{years}{years of 'NHANES' database. If years is missing, nhs_years_web() will be
used to be added.}
}
\value{
add "years.nhanes" file to "nhanesR" directory
}
\description{
Config years of 'NHANES' database
}
