% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_cache.R
\name{.nhanes_cache_subdir}
\alias{.nhanes_cache_subdir}
\title{Return a cache subdirectory path, creating it if needed}
\usage{
.nhanes_cache_subdir(...)
}
\description{
Return a cache subdirectory path, creating it if needed
}
\keyword{internal}
