% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_http.R
\name{.nhanes_check_pkg}
\alias{.nhanes_check_pkg}
\title{Check that a suggested package is installed}
\usage{
.nhanes_check_pkg(pkg, call = rlang::caller_env())
}
\description{
Check that a suggested package is installed
}
\keyword{internal}
