% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.nhanes_download_by_filename}
\alias{.nhanes_download_by_filename}
\title{Download a single NHANES file by its exact catalog filename}
\usage{
.nhanes_download_by_filename(filename, cycle, refresh, add_cycle_col)
}
\description{
Download a single NHANES file by its exact catalog filename
}
\keyword{internal}
