% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_http.R
\name{.nhanes_lmf_url}
\alias{.nhanes_lmf_url}
\title{Build the CDC FTP URL for a public-use LMF .dat file}
\usage{
.nhanes_lmf_url(cycle)
}
\description{
Build the CDC FTP URL for a public-use LMF .dat file
}
\keyword{internal}
