% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{.nhanes_parse_lmf_dat}
\alias{.nhanes_parse_lmf_dat}
\title{Parse a single LMF .dat file using the internal column specification}
\usage{
.nhanes_parse_lmf_dat(dat_path, cycle)
}
\description{
Parse a single LMF .dat file using the internal column specification
}
\keyword{internal}
