% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhis_mortality.R
\name{.nhis_parse_lmf_dat}
\alias{.nhis_parse_lmf_dat}
\title{Parse a single NHIS LMF .dat file using the internal column specification}
\usage{
.nhis_parse_lmf_dat(dat_path, year)
}
\description{
Parse a single NHIS LMF .dat file using the internal column specification
}
\keyword{internal}
