% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_get.R
\name{get_config}
\alias{get_config}
\alias{get_config_path}
\alias{get_config_items}
\alias{get_config_years}
\title{Get config information of "NHANES" database}
\usage{
get_config_path()

get_config_items()

get_config_years(range = TRUE)
}
\arguments{
\item{range}{logical for years}
}
\value{
config information of "NHANES" database
}
\description{
Get config information of "NHANES" database
}
\examples{
\donttest{
get_config_path()
get_config_years()
get_config_items()

}
}
