% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{nhanes_lmf_cycles}
\alias{nhanes_lmf_cycles}
\title{List NHANES cycles with a public-use LMF}
\usage{
nhanes_lmf_cycles()
}
\value{
Character vector of cycle labels.
}
\description{
List NHANES cycles with a public-use LMF
}
\seealso{
\code{\link[=nhanes_mortality_link]{nhanes_mortality_link()}}, \code{\link[=nhanes_mortality_download]{nhanes_mortality_download()}}
}
