% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhis_mortality.R
\name{nhis_lmf_years}
\alias{nhis_lmf_years}
\title{List NHIS survey years with a public-use LMF}
\usage{
nhis_lmf_years()
}
\value{
Character vector of NHIS survey years, \code{"1986"} through \code{"2018"}.
}
\description{
List NHIS survey years with a public-use LMF
}
\seealso{
\code{\link[=nhis_mortality_link]{nhis_mortality_link()}}, \code{\link[=nhis_mortality_download]{nhis_mortality_download()}}
}
