% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_check.R
\name{nhs_check}
\alias{nhs_check}
\title{Check error files
The error file is defined by the absence of the following five formats: xpt or sas7bdat,
tsv, label, codebook and update.}
\usage{
nhs_check(years, items)
}
\arguments{
\item{years}{one or more years}

\item{items}{one or more items}
}
\value{
if error files exist, green text will be print.
}
\description{
Check error files
The error file is defined by the absence of the following five formats: xpt or sas7bdat,
tsv, label, codebook and update.
}
\examples{
\donttest{
nhs_check()
}
}
