% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_codebook.R
\name{nhs_codebook}
\alias{nhs_codebook}
\title{Codebook for variable}
\usage{
nhs_codebook(..., files)
}
\arguments{
\item{...}{one(suggest) or more variable names}

\item{files}{missing(suggest) or file path of nhs_files_pc() or dataframe of nhs_select()}
}
\value{
variable codebook
}
\description{
Codebook for variable
}
