% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_colnames.R
\name{nhs_colnames}
\alias{nhs_colnames}
\alias{nhs_colnames.character}
\alias{nhs_colnames.list}
\title{exact colnames name}
\usage{
nhs_colnames(
  files,
  order = FALSE,
  grep = NULL,
  exact = NULL,
  ex_grep = NULL,
  ex_exact = NULL,
  combine = NULL,
  view = FALSE
)

\method{nhs_colnames}{character}(
  files,
  order = FALSE,
  grep = NULL,
  exact = NULL,
  ex_grep = NULL,
  ex_exact = NULL,
  combine = NULL,
  view = FALSE
)

\method{nhs_colnames}{list}(
  files,
  order = FALSE,
  grep = NULL,
  exact = NULL,
  ex_grep = NULL,
  ex_exact = NULL,
  combine = NULL,
  view = FALSE
)
}
\arguments{
\item{files}{file names}

\item{order}{logical. whether to order by colnames}

\item{grep}{filter colnames by grep() function}

\item{exact}{filter colnames by exact match}

\item{ex_grep}{exclude by grep() function}

\item{ex_exact}{exclude by exact match}

\item{combine}{combine colnames together}

\item{view}{logical. Whether to view}
}
\value{

}
\description{
exact colnames name
}
\examples{
\donttest{
# colnames for demo
demo <- nhs_files_pc(pattern = 'demo',file_ext = 'tsv') |>
    set::grep_not_or('p_demo')

nhs_colnames(files = demo,view = T,
                ex_grep = 'wtirep|wtmrep',
                combine = c('aialang&aialanga','dmdborn&dmdborn2&dmdborn4',
                            'dmdhrage&dmdhragz','dmdhrbrn&dmdhrbr2&dmdhrbr4',
                            'dmdhredu&dmdhredz','dmdhrmar&dmdhrmaz','dmdhsedu&dmdhsedz',
                            'dmqmilit&dmqmiliz','indfminc&indfmin2','indhhinc&indhhin2'))

}
}
