% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_config_path.R
\name{nhs_config_path}
\alias{nhs_config_path}
\title{Config path of 'NHANES' database}
\usage{
nhs_config_path(nhs.path)
}
\arguments{
\item{nhs.path}{path of 'NHANES' database}
}
\value{
change the internal path
}
\description{
Config path of 'NHANES' database
}
