% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_data.R
\name{nhs_data}
\alias{nhs_data}
\alias{nhs_data_web}
\title{available data in NHANES database}
\usage{
nhs_data_web(years, cat = TRUE)
}
\arguments{
\item{years}{one or more years}

\item{cat}{logical, wheter to show the process}
}
\value{
available data
}
\description{
available data in NHANES database
}
\examples{
nhs_data_web(years=1999)
nhs_data_web(years=c(1999,2001))
nhs_data_web(nhs_year(range = F))
}
