% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_download.R
\name{nhs_download}
\alias{nhs_download}
\title{Download data from 'NHANES' database}
\usage{
nhs_download(years, data, files, mode, filetable, cat = TRUE)
}
\arguments{
\item{years}{one or more years}

\item{data}{one or more data}

\item{files}{which files to be download}

\item{mode}{character. The mode with which to write the file. Useful values are "wb" (binary), "w" and "ab".}
}
\value{
download 'NHANES' database data into local computer and keep the same order.
}
\description{
Download data from 'NHANES' database
}
\examples{
\donttest{
nhs_download(years = c(2017,2019),data = c('d','l'))
}
}
