% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_file.R
\name{nhs_file}
\alias{nhs_file}
\alias{nhs_file_web}
\alias{nhs_file_pc}
\alias{nhs_file_pg}
\title{available data in NHANES database}
\usage{
nhs_file_web(years, data, cat = TRUE)

nhs_file_pc(data, years, type, cat = TRUE)

nhs_file_pg(data, years, type, cat = TRUE)
}
\arguments{
\item{years}{one or more years}

\item{cat}{logical, wheter to show the process}
}
\value{
available data
}
\description{
available data in NHANES database
}
\examples{
# nhs_file_web(years=1999,data=c('d','e','l','q'))
# nhs_file_web(years=2019,data=c('d','e','l','q'))
# nhs_file_web(nhs_year(range = F),'demo')
}
