% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_data.R
\name{nhs_items}
\alias{nhs_items}
\alias{nhs_items_web}
\alias{nhs_items_pc}
\title{available items in NHANES database}
\usage{
nhs_items_web(years, cat = TRUE)

nhs_items_pc(years)
}
\arguments{
\item{years}{one or more years}

\item{cat}{logical, wheter to show the process}
}
\value{
available items
}
\description{
available items in NHANES database
}
\examples{
nhs_items_web(years=1999)
nhs_items_web(years=c(1999,2001))
nhs_items_web(nhs_year(range = F))
}
