% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_label.R
\name{nhs_label}
\alias{nhs_label}
\title{label for variable}
\usage{
nhs_label(..., files)
}
\arguments{
\item{...}{one(suggest) or more variable names}

\item{files}{missing(suggest) or file path of nhs_files_pc()}
}
\value{
variable label
}
\description{
label for variable
}
