% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_read.R
\name{nhs_read}
\alias{nhs_read}
\title{Read data from 'NHANES' database in local PC}
\usage{
nhs_read(..., years, exclude = NULL, var.label = FALSE, cookbook = FALSE)
}
\arguments{
\item{...}{one or more data file name.}

\item{years}{one or years}

\item{exclude}{exclude some file}

\item{var.label}{logical, whether to add label for variable}

\item{cookbook}{logical, whether to decode variable}
}
\value{
a list contains dataframe
}
\description{
Read data from 'NHANES' database in local PC
}
\examples{
\donttest{
nhs_read('demo')
}
}
