% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_select.R
\name{nhs_select}
\alias{nhs_select}
\title{Select column for results of nhs_read() function}
\usage{
nhs_select(x, ...)
}
\arguments{
\item{x}{results of nhs_read() function}

\item{...}{one ore more columns}
}
\value{
selected column data
}
\description{
Select column for results of nhs_read() function
}
