% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_variable_pc.R
\name{nhs_variable_pc}
\alias{nhs_variable_pc}
\title{exact variable name}
\usage{
nhs_variable_pc(
  files,
  order = FALSE,
  grep = NULL,
  exact = NULL,
  ex_grep = NULL,
  ex_exact = NULL,
  combine = NULL,
  view = FALSE
)
}
\arguments{
\item{files}{file names}

\item{order}{logical. whether to order by variable}

\item{grep}{filter variable by grep() function}

\item{exact}{filter variable by exact match}

\item{ex_grep}{exclude by grep() function}

\item{ex_exact}{exclude by exact match}

\item{combine}{combine variables together}

\item{view}{logical. Whether to view}
}
\value{

}
\description{
exact variable name
}
\examples{
\donttest{
# variable for demo
demo <- nhs_files_pc(pattern = 'demo',file_ext = 'tsv') |>
    set::grep_not_or('p_demo')

nhs_variable_pc(files = demo,view = T,
                ex_grep = 'wtirep|wtmrep',
                combine = c('aialang&aialanga','dmdborn&dmdborn2&dmdborn4',
                            'dmdhrage&dmdhragz','dmdhrbrn&dmdhrbr2&dmdhrbr4',
                            'dmdhredu&dmdhredz','dmdhrmar&dmdhrmaz','dmdhsedu&dmdhsedz',
                            'dmqmilit&dmqmiliz','indfminc&indfmin2','indhhinc&indhhin2'))

}
}
