% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_view.R
\name{nhs_view}
\alias{nhs_view}
\title{View result of nsh_datafile()}
\usage{
nhs_view(x, ...)
}
\arguments{
\item{x}{result of nsh_datafile()}

\item{...}{highlight}
}
\value{
data in viewer pannel in RStudio
}
\description{
View result of nsh_datafile()
}
