% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_year.R
\name{nhs_year}
\alias{nhs_year}
\alias{nhs_year_web}
\alias{nhs_year_pc}
\alias{nhs_year_pg}
\title{available Year in NHANES database}
\usage{
nhs_year_web(range = TRUE)

nhs_year_pc(range = TRUE)

nhs_year_pg(range = TRUE)
}
\arguments{
\item{range}{whether to show year interval}
}
\value{
available year, two years interval
}
\description{
available Year in NHANES database
}
\examples{
nhs_year_web()
nhs_year_pc()
nhs_year_pg()
}
