% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_items}
\alias{prepare_items}
\title{prepare items from config
ignore capital and little letters and left match}
\usage{
prepare_items(items)
}
\arguments{
\item{items}{one or more items}
}
\value{
items of config
}
\description{
prepare items from config
ignore capital and little letters and left match
}
