% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_years}
\alias{prepare_years}
\title{prepare years from config}
\usage{
prepare_years(years, range = TRUE)
}
\arguments{
\item{years}{one or more years}

\item{range}{logical.}
}
\value{
years of config
}
\description{
prepare years from config
}
