% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_browse.R
\name{nhs_browse}
\alias{nhs_browse}
\title{Browse "NHANES" webpage}
\usage{
nhs_browse(years, data)
}
\arguments{
\item{years}{one or more years}

\item{data}{one or more data}
}
\value{
browse webpage
}
\description{
Browse "NHANES" webpage
}
