% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_file.R
\name{nhs_files}
\alias{nhs_files}
\alias{nhs_files_web}
\alias{nhs_files_pc}
\alias{nhs_files_tsv}
\title{available data in NHANES database}
\usage{
nhs_files_web(years, items, cat = TRUE, withdrawn = FALSE)

nhs_files_pc(items, years, pattern = NULL, file_ext = NULL, cat = TRUE)

nhs_files_tsv(items, years, pattern = NULL, cat = TRUE)
}
\arguments{
\item{years}{one or more years}

\item{cat}{logical, wheter to show the process}

\item{withdrawn}{logical, whether include withdrawned file}

\item{pattern}{for nhs_files_pc()}

\item{file_ext}{file extensions for nhs_files_pc(), default is NULL to list all files}
}
\value{
available data
}
\description{
available data in NHANES database
}
\examples{
# nhs_files_web(years=1999,items=c('d','e','l','q'))
# nhs_files_web(years=2019,items=c('d','e','l','q'))
# nhs_files_web(nhs_year(range = F),'demo')
}
