% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_read.R
\name{nhs_read}
\alias{nhs_read}
\title{Read data from 'NHANES' database in local PC}
\usage{
nhs_read(..., label = FALSE, codebook = FALSE, nrows = Inf)
}
\arguments{
\item{...}{one or more data file path}

\item{label}{logical, whether to add label for variable}

\item{codebook}{logical, whether to decode variable}

\item{nrows}{The maximum number of rows to read.}
}
\value{
a list contains dataframe
}
\description{
Read data from 'NHANES' database in local PC
}
\examples{
\donttest{

demo <- nhs_files_pc(pattern = 'demo',file_ext = 'tsv') |>
        set::grep_not_or('p_demo')
nhs_read(demo)
}
}
