% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_select.R, R/nhs_select.R
\name{nhs_select}
\alias{nhs_select}
\title{Select column for results of nhs_read() function}
\usage{
nhs_select(data, ..., force_bind = FALSE)

nhs_select(data, ..., force_bind = FALSE)
}
\arguments{
\item{data}{result of nhs_read()}

\item{...}{paired names, connected by colon.}

\item{force_bind}{logical.}

\item{x}{results of nhs_read() function}

\item{select}{logical. whether to select the}
}
\value{
selected column data


}
\description{
Select column for results of nhs_read() function

Rename column names for result of nhs_read()
}
