% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_select.R, R/nhs_select.R
\name{nhs_select}
\alias{nhs_select}
\title{Select column for results of nhs_read() function}
\usage{
nhs_select(nhs, exact = NULL, grep = NULL, ex_grep = NULL)

nhs_select(nhs, exact = NULL, grep = NULL, ex_grep = NULL)
}
\arguments{
\item{nhs}{result of nhs_read()}

\item{exact}{exact column names}

\item{grep}{fuzzy match}

\item{ex_grep}{exclude by fuzzy match}

\item{x}{results of nhs_read() function}

\item{...}{one ore more columns}
}
\value{
selected column data

selected list
}
\description{
Select column for results of nhs_read() function

Select column for nhs_read() result
}
