% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_select.R
\name{nhs_select}
\alias{nhs_select}
\title{Rename column names for result of nhs_read()}
\usage{
nhs_select(data, ..., force_bind = FALSE)
}
\arguments{
\item{data}{result of nhs_read()}

\item{...}{paired names, connected by colon.}

\item{force_bind}{logical.}

\item{select}{logical. whether to select the}
}
\value{

}
\description{
Rename column names for result of nhs_read()
}
