% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_update.R
\name{nhs_update}
\alias{nhs_update}
\title{update "NHANES" database}
\usage{
nhs_update(years, data, download = TRUE, redown = TRUE)
}
\arguments{
\item{years}{one or more years}

\item{data}{one or more data}

\item{download}{logical. Whether to download}
}
\value{
update local "NHANES" database
}
\description{
update "NHANES" database
}
\examples{
\donttest{
nhs_update()
}
}
