% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_year.R
\name{nhs_year}
\alias{nhs_year}
\alias{nhs_years_web}
\title{available Year in NHANES database}
\usage{
nhs_years_web(range = TRUE)
}
\arguments{
\item{range}{whether to show year interval}
}
\value{
available year, two years interval
}
\description{
available Year in NHANES database
}
\examples{
nhs_years_web()
}
